move LazyEncoder into hqwebapp
